dali-doc-gen/requirements.txt
GitPython>=3.1.30
aiohttp>=3.8.0
Jinja2>=3.1.2
PyYAML>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.30.0
